Fixed GameTimer to be 32 bits.

+++ TRUtil/LaterClassicGameData.cs
@@ -16,7 +16,7 @@ public abstract class LaterClassicGameData : BaseGameData
     public static MemoryWatcher<uint> GfLevelComplete => (MemoryWatcher<uint>)Watchers?["GfLevelComplete"];
 
     /// <summary>Represents a "global" timer tracking ticks since New Game. It is saved/reloaded. </summary>
-    public static MemoryWatcher<ulong> GameTimer => (MemoryWatcher<ulong>)Watchers?["GameTimer"];
+    public static MemoryWatcher<uint> GameTimer => (MemoryWatcher<uint>)Watchers?["GameTimer"];
 
     /// <summary>Indicates if a loading screen is active.</summary>
     public static MemoryWatcher<bool> Loading => (MemoryWatcher<bool>)Watchers?["Loading"];
